Change for loops to be in parallel
Thought this might be a good change so it affects each object at the exact same time in parallel instead of running through the list. I'm not positive this even compiles since I'm writing it in GitHub. Might need some import. and definitely requires java 8 or higher, so it won't work on lower versions.