Implemented IDs, player object, and movement.
I created a bunch of individual enum IDs to kind of give a sense of what I want to do. Feel free to figure out some of the weapons types or enemies that I've added in there. For example, a moving boomerang that comes back to you, an arrow from a bow, the arc of the sword, etc. I'm going to do a bit more player stuff and then get add the map functionality.

On a side note, the game constructor is a bit of a test area currently so feel free to add your objects in there. Eventually each location/level will be it's own class and will simply be controlled by the handler/exit values. The game constructor will just play the start screen sequence and direct toward the saved state on selection.