Removed multiple key movement bug.
I know the new code is a bit verbose and a lot of the extra steps seem redundant, but the controls would glitch out with anything less whenever I would press the keys quickly/hold other keys down. I finally got it to the point where no matter what I did concerning WASD, The program would still execute it exactly how I wanted it to. It feels really crisp and I tend to hold down multiple keys when kiting or just moving around in general so the change needed to be made.