Added map, camera, and collision functionality.
The collision method can be improved, but I just added the quick and dirty method that was illustrated in the tut. Eventually I'd like it to be pixel perfect. I also included two different camera formulas for a classic fixed center, locked on edges style and a more modern "flowing/smooth" camera. The Parallel stream for rendering crashes the program so we'll need to either address the problem or go with the less optimal code. The test map is in a resources folder for the project build path so the code shouldn't need any changes if you set it up the same way on in your IDE.